refactor: move smt decorator handlers into stand-alone module
